requirements.txt
fastapi
uvicorn
pytorch
matplotlib
numpy
scikit-learn